User info and change passowrd end point, was added to the project

+++ TodoApp/routers/todos.py
@@ -22,23 +22,22 @@ def get_db():
         db.close()
         
 db_dependency = Annotated[Session, Depends(get_db)]
+
 user_dependency = Annotated[dict, Depends(get_current_user)]
     
         
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(user: user_dependency, db: db_dependency):
-    return db.query(Todos).filter(Todos.owner_id == user.get("user_id")).all()
-
-
-@router.get("/all_todos", status_code=status.HTTP_200_OK)
-async def read_completed(db: db_dependency):
-    return db.query(Todos).all()
-
+async def read_all(user: user_dependency,
+                   db: db_dependency):
+    return db.query(Todos).filter(Todos.owner_id == user.get("id")).all()
 
 
 @router.get("/todos/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db : db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency,
+                    db : db_dependency, 
+                    todo_id: int = Path(gt=0)):
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model:
         return todo_model
     raise HTTPException(status_code=404, detail="Todo not found")
@@ -51,16 +50,18 @@ async def create_todo(user: user_dependency,
     
     if user is None:
         raise HTTPException(status_code=401, detail="Authentication Failed")
-    todo_model = Todos(**todo_request.dict(), owner_id=user.get("user_id"))
+    todo_model = Todos(**todo_request.dict(), owner_id=user.get("id"))
     db.add(todo_model)
     db.commit()
     return todo_model
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, 
+async def update_todo(user: user_dependency,
+                      db: db_dependency, 
                       todo_request: TodoRequest,
                       todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.id == user.get('id')).first()
     if todo_model:
         todo_model.title = todo_request.title
         todo_model.description = todo_request.description
@@ -73,8 +74,12 @@ async def update_todo(db: db_dependency,
     raise HTTPException(status_code=404, detail="Todo not found")
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency,
+                      db: db_dependency, 
+                      todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.id == user.get('id')).first()
     if todo_model:
         db.delete(todo_model)
         db.commit()
